split interface and mixin into two files
